Perform inflation check on mirror node
Each time we check inflation for a node, the gettxoutsetinfo call blocks all other RPC calls to that node. This interferes with regular polling, so we can only do inflation checks infrequently.

Moving the inflation check and gettxoutsetinfo to the mirror node allows us to check inflation on more nodes and more frequently.