Poll mirror during rollback, and more detailed logging
Invalidating a block can cause the node to switch the chaintip to a blocks that we don't have in our database. The loop over block.children would then not achieve a rollback.

If the tip doesn't move after on iteration, poll the node to fetch new blocks.

Give up after 100 attempts to prevent an infinite loop.