Track which node marks a block valid / invalid
Previously we would send an invalid block alert if:
1. One node considered a chaintip invalid
2. The block exists in our database
This implicitly assumed that each blocks in our database came from a node that considered that block to be valid.

This commit removes that assumption. We now store every fully processed block we see: active chaintip, valid fork and invalid, but not yet headers-only.

We each block we track which nodes consider it valid and which consider it invalid. This data may be incomplete, but it's at least as good as before. An alert is triggered for any block were at least one node considers it valid and one node considers it invalid.

One practical benefit of this change is that we can make an RSS feed for all invalid blocks, rather than just conflicting invalid blocks.