Suppress stale block alerts if there is a validity conflict
If two nodes disagree on consensus rules, e.g. after a hardfork, there will be two chains. This causes a stream of stale block alerts.

The most elegant solution to that is to check each stale block and see if it descends from a the same same of a fork.

An easier solution is to just suppress stale block alerts if there was a consensus conflict. Eventually one of the nodes is deleted from the site and the alert is dismissed. At that point we resume notifications.