Use non-blocking poll for stale candidate info
Cache for stale candidate info is primed by the rake heavy task. When a user visits the page before this happens, it will try to generate the resul on the fly. This can take a very long time, result in a timeout, and be problematic if a user refreshes the page.

Instead, we now return a empty result and ask the client to try again after a few seconds. Once the background task is finished, the page will update.