Ignore prune status when fetching transactions
When fetching transactions we either the use the node that first saw it, or we use the most recent Bitcoin Core node (which is not pruned).

In the first case it's possible the node may have pruned the transaction, so we mark it as such.  Next time it will be tried with a new node.

In the second case, if this is a new block then perhaps the latest node hasn't seen it yet. We just try again later.